Throttling (#17)
* Add `ping` and `version` endpoints

* Enable request throttling

* Get the system's available disk space

* 'filters' ~> 'transformers'

* Implement a size restriction per-user

* Keep websockets alive with a ping every ten seconds

* Better websocket error reporting

* v0.5.0

* Show the server's reported version on the client

* Share relevant storage stats with the client after each write

* Update the client's cache of user data stats after each relevant write

* Prettier usage indicator

+++ server/db.ts
@@ -1,12 +1,14 @@
 import type { DataItem, Unsubscribe, UserKeys } from "./database/index.js";
 import type { DocUpdate } from "./database/io.js";
 import type { Request } from "express";
-import type { WebSocket } from "ws";
+import type { WebSocket } from "./database/websockets.js";
 import { asyncWrapper } from "./asyncWrapper.js";
 import { BadRequestError, NotFoundError, respondData, respondSuccess } from "./responses.js";
+import { close, send, WebSocketCode } from "./database/websockets.js";
 import { handleErrors } from "./handleErrors.js";
 import { ownersOnly, requireAuth } from "./auth/index.js";
 import { Router } from "express";
+import { statsForUser } from "./database/io.js";
 import {
 	CollectionReference,
 	DocumentReference,
@@ -53,10 +55,28 @@ function webSocket(ws: WebSocket, req: Request<Params>): void {
 	const uid = (req.params.uid ?? "") || null;
 	const collectionId = (req.params.collectionId ?? "") || null;
 	const documentId = (req.params.documentId ?? "") || null;
-	if (uid === null || collectionId === null || !isCollectionId(collectionId)) {
-		ws.send(JSON.stringify({ message: "No data found" }));
-		return ws.close();
-	}
+
+	// Ensure valid input
+	if (uid === null) return close(ws, WebSocketCode.PROTOCOL_ERROR, "Missing user ID");
+	if (collectionId === null)
+		return close(ws, WebSocketCode.PROTOCOL_ERROR, "Missing collection ID");
+	if (!isCollectionId(collectionId))
+		return close(ws, WebSocketCode.PROTOCOL_ERROR, "Invalid collection ID");
+
+	// Send an occasional ping
+	// If the client doesn't respond a few times consecutively, assume they aren't coming back
+	let timesNotThere = 0;
+	const pingInterval = setInterval(() => {
+		if (timesNotThere > 5) {
+			process.stdout.write("Client didn't respond after 5 tries. Closing\n");
+			close(ws, WebSocketCode.WENT_AWAY, "Client did not respond to pings, probably dead");
+			clearInterval(pingInterval);
+			return;
+		}
+		send(ws, "ARE_YOU_STILL_THERE");
+		timesNotThere += 1; // this goes away if the client responds
+	}, 10000); // 10 second interval
+
 	const collection = new CollectionReference<UserKeys>(uid, collectionId);
 	let unsubscribe: Unsubscribe;
 
@@ -66,40 +86,42 @@ function webSocket(ws: WebSocket, req: Request<Params>): void {
 		const ref = new DocumentReference(collection, documentId);
 		unsubscribe = watchUpdatesToDocument(ref, data => {
 			console.debug(`Got update for document at ${ref.path}`);
-			ws.send(
-				JSON.stringify({
-					message: "Here's your data",
-					dataType: "single",
-					data,
-				})
-			);
+			send(ws, {
+				message: "Here's your data",
+				dataType: "single",
+				data,
+			});
 		});
 	} else {
 		unsubscribe = watchUpdatesToCollection(collection, data => {
 			console.debug(`Got update for collection at ${collection.path}`);
-			ws.send(
-				JSON.stringify({
-					message: "Here's your data",
-					dataType: "multiple",
-					data,
-				})
-			);
+			send(ws, {
+				message: "Here's your data",
+				dataType: "multiple",
+				data,
+			});
 		});
 	}
 
 	ws.on("message", msg => {
 		try {
-			const uid = (req.params.uid ?? "") || null;
-			if ((msg as Buffer).toString() === "STOP") {
-				process.stdout.write("Received STOP\n");
-				unsubscribe();
-				return ws.close();
+			const message = (msg as Buffer).toString();
+
+			// ** Stuff we expect to hear from the client:
+			switch (message) {
+				case "START": // nop
+				case "YES_IM_STILL_HERE":
+					timesNotThere = 0;
+					return;
+				case "STOP":
+					unsubscribe();
+					return close(ws, WebSocketCode.NORMAL, "Received STOP message from client");
+				default:
+					return close(ws, WebSocketCode.PROTOCOL_ERROR, "Received unknown message from client");
 			}
-			process.stdout.write(
-				`User '${uid ?? "null"}' sent message: '${JSON.stringify(msg.valueOf())}'\n`
-			);
 		} catch (error: unknown) {
 			console.error(error);
+			close(ws, WebSocketCode.PROTOCOL_ERROR, "Couldn't process that");
 		}
 	});
 }
@@ -144,7 +166,10 @@ export function db(this: void): Router {
 				if (!isArrayOf(providedData, isDocumentWriteBatch)) throw new BadRequestError();
 
 				// Ignore an empty batch
-				if (!isNonEmptyArray(providedData)) return respondSuccess(res);
+				if (!isNonEmptyArray(providedData)) {
+					const { totalSpace, usedSpace } = await statsForUser(uid);
+					return respondSuccess(res, { totalSpace, usedSpace });
+				}
 
 				// Separate delete and set operations
 				const setOperations: Array<DocUpdate<DataItem>> = [];
@@ -173,7 +198,8 @@ export function db(this: void): Router {
 					await deleteDocuments(deleteOperations);
 				}
 
-				respondSuccess(res);
+				const { totalSpace, usedSpace } = await statsForUser(uid);
+				respondSuccess(res, { totalSpace, usedSpace });
 			})
 		)
 		.post<Params>(
@@ -190,27 +216,36 @@ export function db(this: void): Router {
 				if (!ref) throw new NotFoundError();
 
 				await setDocument(ref, providedData);
-				respondSuccess(res);
+				const { totalSpace, usedSpace } = await statsForUser(uid);
+				respondSuccess(res, { totalSpace, usedSpace });
 			})
 		)
 		.delete<Params>(
 			"/users/:uid/:collectionId",
 			asyncWrapper(async (req, res) => {
+				const uid = (req.params.uid ?? "") || null;
+				if (uid === null) throw new NotFoundError();
+
 				const ref = collectionRef(req);
 				if (!ref) throw new NotFoundError();
 
 				await deleteCollection(ref);
-				respondSuccess(res);
+				const { totalSpace, usedSpace } = await statsForUser(uid);
+				respondSuccess(res, { totalSpace, usedSpace });
 			})
 		)
 		.delete<Params>(
 			"/users/:uid/:collectionId/:documentId",
 			asyncWrapper(async (req, res) => {
+				const uid = (req.params.uid ?? "") || null;
+				if (uid === null) throw new NotFoundError();
+
 				const ref = documentRef(req);
 				if (!ref) throw new NotFoundError();
 
 				await deleteDocument(ref);
-				respondSuccess(res);
+				const { totalSpace, usedSpace } = await statsForUser(uid);
+				respondSuccess(res, { totalSpace, usedSpace });
 			})
 		)
 		.use(handleErrors);
